updated buttoms and placements
Soon done need logic
